feat(invites)!: implement employee invitation endpoint
Adds the functionality to invite employees to a yard. A new endpoint
`POST /v2/yards/{yardId}/invites` is created to handle the creation
of new employee invitations.

The response for the `GET /v2/yards/{yardId}` endpoint has also been
updated to return a structured list of invites.

BREAKING CHANGE: The `GET /v2/yards/{yardId}` endpoint now returns a
different structure for the `invites` collection. The items in the
collection are now mapped to a dedicated response object instead of the
internal database model.