refactor(db)!: simplify vehicle entity by removing join table
Removes the `YardVehicle` join table and merges its properties directly into the `Vehicle` entity. This simplifies the data model, changing the relationship between `Yard` and `Vehicle` to a direct one-to-many.

A new relationship between `Vehicle` and `YardEmployee` has been added to track the assignee. The vehicle API endpoints are temporarily disabled as they are broken by this change and will be updated in a future commit.

BREAKING CHANGE: The 'YardVehicle' join table has been removed. Vehicle status and yard relationship data are now stored directly on the 'vehicles' table.